WIP: moving color models into separate files to keep the code readable (and testable).
